requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
asttokens==2.0.5
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
boto3==1.20.23
botocore==1.23.23
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.9
django-bootstrap4==21.1
django-environ==0.8.1
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.12.4
executing==0.8.2
frozenlist==1.2.0
gunicorn==20.1.0
icecream==2.1.1
idna==3.3
jmespath==0.10.0
multidict==5.2.0
Pillow==9.0.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
Pygments==2.10.0
python-dateutil==2.8.2
python-http-client==3.3.5
pytz==2021.3
qrcode==7.3.1
requests==2.26.0
s3transfer==0.5.0
sendgrid==6.9.5
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
telepot==12.7
typing_extensions==4.0.1
urllib3==1.26.7
whitenoise==5.3.0
yarl==1.7.2
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.4.1
asttokens==2.0.5
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
boto3==1.20.23
botocore==1.23.23
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-bootstrap4==21.1
django-environ==0.8.1
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.12.4
executing==0.8.2
frozenlist==1.2.0
gunicorn==20.1.0
icecream==2.1.1
idna==3.3
jmespath==0.10.0
kombu==5.2.4
multidict==5.2.0
packaging==21.3
Pillow==9.0.0
prompt-toolkit==3.0.31
psycopg2==2.9.2
psycopg2-binary==2.9.2
Pygments==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-http-client==3.3.5
pytz==2021.3
qrcode==7.3.1
redis==4.3.4
requests==2.26.0
s3transfer==0.5.0
sendgrid==6.9.5
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
telepot==12.7
typing_extensions==4.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.14.1
yarl==1.7.2
